Upgrade the icon & banner
