+++ app/build.gradle
@@ -16,8 +16,8 @@ android {
         applicationId 'com.itssuryansh.taaveez'
         minSdk 16
         targetSdk 33
-        versionCode 2
-        versionName "v1.1.2"
+        versionCode 3
+        versionName "v1.1.3"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
     }
