+++ app/src/main/java/com/itssuryansh/taaveez/TaaveezEntity.kt
@@ -16,5 +16,4 @@ data class TaaveezEntity(
     val favorite:Boolean=false,
     val image : String?=null
 
-
 )
