+++ app/src/main/java/com/itssuryansh/taaveez/database/TaaveezDatabase.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.itssuryansh.taaveez.TaaveezDao
 import com.itssuryansh.taaveez.TaaveezEntity
 
 @Database(entities = [TaaveezEntity::class],version=3)
