+++ app/src/main/java/com/itssuryansh/taaveez/helper/Constants.kt
@@ -1,0 +1,17 @@
+package com.itssuryansh.taaveez.helper
+
+object Constants {
+
+    const val TAAVEEZ_CONTENT_TOPIC: String = "taaveez_content_topic"
+    const val TAAVEEZ_CONTENT_DESCRIPTION : String = "taaveez_content_description"
+    const val CREATED_DATE : String ="created_date"
+    const val UPDATED_DATE : String ="updated_date"
+    const val MyPrefs : String = "MyPrefss"
+    const val introShown : String = "introShown"
+    const val LINK : String = "link"
+    const val ID = "id"
+    const val IS_CONTENT_COMPLETE_STATUS : String = "is_content_complete_status"
+    var Theme : String = "0"
+    const val IS_DETAILED_VIEW = "myBooleanValue" // Key for the boolean value
+
+}
