+++ app/src/main/res/layout/menu_header.xml
@@ -30,6 +30,7 @@
             android:layout_height="wrap_content"
             android:text="@string/app_name"
             android:textSize="30sp"
+            android:layout_marginTop="5dp"
             android:textStyle="bold"
             android:textColor="@color/white"
             app:layout_constraintEnd_toEndOf="parent"
