+++ app/src/main/res/layout/one_time_intro_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".activity.OneTimeIntroActivity">
+    tools:context=".activity.firstscreen.OneTimeIntroActivity">
 
 
     <FrameLayout
