+++ app/src/main/res/values-hi-rIN/strings.xml
@@ -21,7 +21,7 @@
     <string name="yes">हां</string>
     <string name="no">नहीं</string>
     <string name="delete_record">रिकॉर्ड हटाएं</string>
-    <string name="appVersion">v1.1.2</string>
+    <string name="appVersion">v1.1.3</string>
     <string name="share_subject">हे, जाओ और इस शानदार एप्लिकेशन को देखें। डाउनलोड करने के लिए इस लिंक पर जाएं। \n\n
 
         यहां दबाएं: https://github.com/Suryansh1720001/Taaveez#taaveez\n\n
