+++ app/src/main/res/values/strings.xml
@@ -20,7 +20,7 @@
     <string name="yes">Yes</string>
     <string name="no">No</string>
     <string name="delete_record">Delete Record</string>
-    <string name="appVersion">v1.1.2</string>
+    <string name="appVersion">v1.1.3</string>
     <string name="share_subject">Hey, Go and check out this cool application. Go to this link to download.\n\n
 
         Link: https://github.com/Suryansh1720001/Taaveez#taaveez\n\n
